move xgboost.explainer module out into top level of package
