+++ src/tabular_trees/xgboost/__init__.py
@@ -1,3 +1,3 @@
 """Module containing functionality for dealing with xgboost models."""
 
-from . import explainer, parser, trees, validate
+from . import parser, trees, validate
