+++ tests/xgboost/test_calculate_shapley_values.py
@@ -3,8 +3,8 @@ import pandas as pd
 import pytest
 import xgboost as xgb
 
+from tabular_trees.explainer import calculate_shapley_values
 from tabular_trees.trees import export_tree_data
-from tabular_trees.xgboost.explainer import calculate_shapley_values
 
 
 @pytest.fixture()
